[56] - Refactor login process and configuration
