+++ src/Auctionify.API/appsettings.Development.json
@@ -10,8 +10,8 @@
   },
   "AuthSettings": {
     "Key": "This is the key used to sign the token",
-    "Issuer": "https://localhost:7045",
-    "Audience": "https://localhost:7045",
+    "Issuer": "https://localhost:7230",
+    "Audience": "https://localhost:7230",
     "AccessTokenExpirationInMinutes": 10,
     "RefreshTokenExpirationInDays": 60
   }
